remove PostTask,  extract the child component PostTaskForm  and rename it as PostTask
