+++ src/pages/PostTask/PostTask.module.scss
@@ -5,7 +5,6 @@
   height: 100%;
   max-width: 30em;
   min-height: 25rem;
-  background-color: #f6f8fd;
   position: relative;
 }
 
@@ -163,11 +162,10 @@
 }
 
 .container {
-  line-height: 1.4;
+  // line-height: 1.4;
   font-family: 'Raleway', sans-serif;
-  width: 480px;
-  height: 550px;
-  background-color: #f6f8fd;
-  border-radius: 0.8em;
-
+  // // width: 480px;
+  // // height: 550px;
+  // // background-color: #f6f8fd;
+  // border-radius: 0.8em;
 }
