--- src/pages/PostTask/PostTaskForm/Welcome/index.js
@@ -1,1 +1,0 @@
-export { default } from './FirstScreen';
