--- src/pages/PostTask/PostTaskForm/index.js
@@ -1,1 +1,0 @@
-export { default } from './PostTaskForm';
