--- src/pages/PostTask/PostTaskStyles.module.scss
@@ -1,9 +1,0 @@
-.container {
-  line-height: 1.4;
-  font-family: 'Raleway', sans-serif;
-  width: 480px;
-  height: 550px;
-  background-color: #f6f8fd;
-  border-radius: 0.8em;
-
-}
