--- src/pages/PostTask/PostTaskTop/PostTaskTop.jsx
@@ -1,17 +1,0 @@
-import React from 'react';
-
-import styles from './PostTaskTop.module.scss';
-
-import TopBar from './TopBar';
-
-function PostTaskTop (props) {
-  return (
-    <div className={styles.Main}>
-      <TopBar>
-        {props.children}
-      </TopBar>
-    </div>
-  )
-}
-
-export default PostTaskTop;
