--- src/pages/PostTask/PostTaskTop/TopBar/index.js
@@ -1,1 +1,0 @@
-export { default } from './TopBar';
