--- src/pages/PostTask/PostTaskTop/index.js
@@ -1,1 +1,0 @@
-export { default } from './PostTaskTop';
