+++ src/pages/PostTask/TaskBudget/BudgetDisplay/BudgetDisplay.jsx
@@ -1,6 +1,6 @@
 import React from 'react'
 
-import styles from '../../PostTaskForm.module.scss'
+import styles from '../../PostTask.module.scss'
 
 function BudgetDisplay({
   taskBudget,
