+++ src/pages/PostTask/TaskDescription/JobDetailsInput/JobDetailsInput.jsx
@@ -1,8 +1,8 @@
 import React from 'react'
 
-import styles from '../../PostTaskForm.module.scss'
+import styles from '../../PostTask.module.scss'
 
-import TextArea from '../../../../../components/TextArea';
+import TextArea from '../../../../components/TextArea';
 
 export default function JobDetailsInput({
   jobDetails,
