+++ src/pages/PostTask/Welcome/WelcomeImg/WelcomeImg.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import styles from '../../PostTaskForm.module.scss';
+import styles from '../../PostTask.module.scss';
 
 import png from './welcome.png';
