+++ src/pages/PostTask/Welcome/WelcomeImg/index.js
