+++ src/pages/PostTask/Welcome/WelcomeMsg/WelcomeMsg.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import styles from '../../PostTaskForm.module.scss';
+import styles from '../../PostTask.module.scss';
 
 function WelcomeMsg() {
   return (
